pyproject.toml
[tool.poetry]
name = "transcript-memory-engine"
version = "0.1.0"
description = "A RAG-based system for querying and analyzing transcripts using local LLMs."
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "transcript_engine" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
httpx = "^0.27.0"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
chromadb = "^0.4.24"
sentence-transformers = "^2.5.1"
ollama = "^0.1.7"
python-dotenv = "^1.0.1"
langchain-text-splitters = "^0.2.1"
numpy = "<2"
streamlit = "^1.45.0"
jinja2 = "^3.1.6"
python-multipart = "^0.0.20"
markdown-it-py = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
